% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abMat.R
\name{abMat}
\alias{abMat}
\title{Generate a species by site abundance matrix}
\usage{
abMat(x, site_id, species_id, fpc = NULL, abundance = NULL)
}
\arguments{
\item{x}{dataframe of individual records}

\item{site_id}{column name string of site IDs}

\item{species_id}{column name string of species names}

\item{fpc}{optional column name string of sampling weights of each record,
between 0 and 1}

\item{abundance}{optional column name string with an alternative abundance
measure such as biomass, canopy cover, body length}
}
\value{
dataframe of species abundances (columns) per site (rows)
}
\description{
Generate a species by site abundance matrix
}
\examples{
x <- data.frame(site_id = rep(c("A", "B", "C"), each = 3), 
  species_id = sample(c("a", "b", "c", "d"), 9, replace = TRUE), 
  fpc = rep(c(0.5, 0.6, 1), each = 3), 
  abundance = seq(1:9))
abMat(x, "site_id", "species_id")
abMat(x, "site_id", "species_id", "fpc")
abMat(x, "site_id", "species_id", "fpc", "abundance")

}
