% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binwFD.R
\name{binwFD}
\alias{binwFD}
\title{Estimate optimal binwidth for histogram according to Freedman-Diaconis rule}
\usage{
binwFD(x)
}
\arguments{
\item{x}{numeric vector to be plotted on histogram}
}
\value{
numeric vector
}
\description{
Estimate optimal binwidth for histogram according to Freedman-Diaconis rule
}
\examples{
x <- rnorm(100)
binwFD(x)

}
