% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitChange.R
\name{bitChange}
\alias{bitChange}
\title{Detect when the values in an ordered vector change}
\usage{
bitChange(x)
}
\arguments{
\item{x}{vector}
}
\value{
integer vector where the value increases each time the value in
\code{x} changes.
}
\description{
Detect when the values in an ordered vector change
}
\examples{
x <- seq_len(5)
all(bitChange(x) == x)
x <- sample(seq_len(5), 20, replace = TRUE)
bitChange(x)

}
