% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUC.R
\name{calcAUC}
\alias{calcAUC}
\title{Calculate Area Under Curve (AUC) using trapesium approximation}
\usage{
calcAUC(x, y)
}
\arguments{
\item{x}{vector of x values}

\item{y}{vector of y values}
}
\description{
Calculate Area Under Curve (AUC) using trapesium approximation
}
