% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkSplit.R
\name{chunkSplit}
\alias{chunkSplit}
\title{Split a vector into chunks of given size}
\usage{
chunkSplit(x, n)
}
\arguments{
\item{x}{vector to split}

\item{n}{size of chunks}
}
\value{
list of length \code{n}
}
\description{
Split a vector into chunks of given size
}
\examples{
x <- seq_len(64)
n <- 8
chunkSplit(x,n)

}
