% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffVar.R
\name{coeffVar}
\alias{coeffVar}
\title{Calculate the coefficient of variation}
\usage{
coeffVar(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
coefficient of variation of vector
}
\description{
Calculate the coefficient of variation
}
\examples{
coeffVar(seq(0, 10))
}
