% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrPlot.R
\name{corrPlot}
\alias{corrPlot}
\title{Create a correlation plot (correlogram)}
\usage{
corrPlot(x, col = c("blue", "white", "red"), ...)
}
\arguments{
\item{x}{dataframe with all numeric columns}

\item{col}{vector of length 3 with character strings of colours}

\item{...}{extra arguments passed to \code{psych::corr.test()}}
}
\value{
ggplot object
}
\description{
Create a correlation plot (correlogram)
}
