% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decDMS.R
\name{decDMS}
\alias{decDMS}
\title{Convert decimal degrees to degrees, minutes, seconds}
\usage{
decDMS(dec, dire = NULL, dataframe = FALSE)
}
\arguments{
\item{dec}{vector of decimal degrees}

\item{dire}{one of "lat" (latitude) or "lon" (longitude) or NULL}

\item{dataframe}{logical, should a dataframe be returned instead of a
formatted string}
}
\value{
character vector of coordinates formatted as degrees, minutes and seconds
}
\description{
Convert decimal degrees to degrees, minutes, seconds
}
\examples{
x <- c(15.53, -15.54, 0.1)
decDMS(x)
decDMS(x, "lon")
decDMS(x, "lat")
decDMS(x, "lat", dataframe = TRUE)

}
