% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmsDeg.R
\name{dmsDeg}
\alias{dmsDeg}
\title{Convert degrees, minutes, seconds to decimal degrees}
\usage{
dmsDeg(degr, minu, seco, dire)
}
\arguments{
\item{degr}{vector of degrees}

\item{minu}{vector of minutes}

\item{seco}{vector of seconds}

\item{dire}{direction, e.g. N/S, E/W}
}
\value{
atomic vector of decimal degrees
}
\description{
Convert degrees, minutes, seconds to decimal degrees
}
\examples{
x <- "15-05-23-S"
dms_split <- strsplit(x, "-")[[1]]
dmsDeg(
  as.numeric(dms_split[1]), 
  as.numeric(dms_split[2]), 
  as.numeric(dms_split[3]), 
  dms_split[4])

}
