% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endPoints.R
\name{endPoints}
\alias{endPoints}
\title{Calculate end points of line given start point, slope, and distance}
\usage{
endPoints(x, y, m, d)
}
\arguments{
\item{x}{x coordinate of start point}

\item{y}{y coordinate of start point}

\item{m}{slope}

\item{d}{distance}
}
\value{
dataframe of coordinates for both endpoints
}
\description{
Calculate end points of line given start point, slope, and distance
}
