% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enl.R
\name{enl}
\alias{enl}
\title{Effective number of layers in a point cloud distribution}
\usage{
enl(x, binwidth)
}
\arguments{
\item{x}{vector of Z (elevation) coordinates}

\item{binwidth}{width of vertical bins in units of x}
}
\value{
atomic vector of length one describing the effective number of layers
in the canopy
}
\description{
Effective number of layers in a point cloud distribution
}
\details{
Uses the Shannon diversity index (Entropy) to estimate the
"Effective Number of Layers" in the vertical profile of a point cloud
distribution.
}
\examples{
x <- rnorm(10000)
enl(x)

}
\references{
Martin Ehbrecht, Peter Schall, Julia Juchheim, Christian Ammer, &
Dominik Seidel (2016). Effective number of layers: A new measure for
quantifying three-dimensional stand structure based on sampling with
terrestrial LiDARForest Ecology and Management, 380, 212–223.
}
