% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPSG2CRS.R
\name{EPSG2CRS}
\alias{EPSG2CRS}
\title{Get a valid CRS object from an EPSG code.}
\usage{
EPSG2CRS(x, wkt = FALSE)
}
\arguments{
\item{x}{vector of EPSG codes.
See http://www.epsg-registry.org/ or http://spatialreference.org/.}

\item{wkt}{logical, if true returns sf WKT CRS object rather than proj4string}
}
\value{
A vector of CRS objects or list of WKT objects
}
\description{
Convert an EPSG code to proj4string or WKT (Well-Known Text)
}
