% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRbind.R
\name{fastRbind}
\alias{fastRbind}
\title{Fast rbind for dataframes}
\usage{
fastRbind(x)
}
\arguments{
\item{x}{list of dataframe objects}
}
\value{
dataframe
}
\description{
Fast rbind for dataframes
}
\examples{
x <- list(
  data.frame(x = seq(1, 5), y = seq(5, 9)),
  data.frame(x = seq(10, 50), y = seq(50, 90)))
fastRbind(x)

}
