% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGenus.R
\name{fillGenus}
\alias{fillGenus}
\title{In a vector of species genera, expand sequential abbreviations}
\usage{
fillGenus(x, abbrev = ".")
}
\arguments{
\item{x}{vector of species genera strings}

\item{abbrev}{character used to signify start of abbreviation}
}
\value{
character vector of expanded species genera
}
\description{
In a vector of species genera, expand sequential abbreviations
}
\examples{
vec <- c("Ochna", "O.", "Ochna", "Brachystegia", "B.", "B.")
fillGenus(vec)
}
