% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks.R
\name{findPeaks}
\alias{findPeaks}
\title{Find local peaks in a distribution}
\usage{
findPeaks(x, m = 3)
}
\arguments{
\item{x}{vector of value series}

\item{m}{number of adjacent points needed to be lower to count as a peak}
}
\value{
vector of positions in x where there is a peak
}
\description{
Find local peaks in a distribution
}
\details{
use findPeaks(-x) to find troughs
}
\examples{
x <- w <- rnorm(n = 100, mean = 0, sd = 1) 
for (t in 2:100) { x[t] <- x[t - 1] + w[t] }
p3 <- findPeaks(x)
p10 <- findPeaks(x, 10)
t10 <- findPeaks(-x, 10)
plot(x, type = "l")
abline(v = p10, col = "green", lwd = 5)
abline(v = p3, col = "blue")
abline(v = t10, col = "red")

}
