% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotLims.R
\name{ggplotLims}
\alias{ggplotLims}
\title{Get axis limits from a ggplot2 plot object}
\usage{
ggplotLims(x)
}
\arguments{
\item{x}{ggplot2 plot object}
}
\value{
named vector of plot axis limits
}
\description{
Get axis limits from a ggplot2 plot object
}
\examples{
library(ggplot2)
dat <- data.frame(
  x = rnorm(1000),
  y = rnorm(1000))
p <- ggplot(dat, aes(x = x, y = y )) +
  geom_point()
ggplotLims(p)

}
