% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexName.R
\name{hexName}
\alias{hexName}
\title{Find the closest named colour to a given six character hex colour}
\usage{
hexName(x, method = "rgb", metric = "euclidean")
}
\arguments{
\item{x}{vector of six digit hex colours}

\item{method}{colour space, either "rgb" or "hsv".}

\item{metric}{distance metric, either "euclidean" or "manhattan".}
}
\value{
vector of nearest named colour
}
\description{
Find the closest named colour to a given six character hex colour
}
\details{
If metric is "euclidean", distances are root sum-of-squares
differences. "manhattan" distances are the sum of absolute differences.
The named colours come from the list of 657 named colours stored in R,
accessible using the \code{colors()} function.
}
\examples{
hexName(c("#117733", "#b58900", "#855C75"))

}
