% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexTrip.R
\name{hexTrip}
\alias{hexTrip}
\title{Find the closest hex triplet to a given six character hex colour}
\usage{
hexTrip(x)
}
\arguments{
\item{x}{vector of six digit hex colours}
}
\value{
vector of nearest hex colour represented as a triplet
}
\description{
Find the closest hex triplet to a given six character hex colour
}
\examples{
hexTrip(c("#117733", "#b58900", "#855C75"))
hexTrip("#fffffff")
hexTrip("855C75")

}
