% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboDataGet.R
\name{koboDataGet}
\alias{koboDataGet}
\title{Get KoboToolbox form data as a dataframe for a given form}
\usage{
koboDataGet(server_url, formid, token = NULL, user = NULL, pass = NULL)
}
\arguments{
\item{server_url}{URL of the KoboToolbox server,
e.g. \url{https://kc.kobotoolbox.org/}}

\item{formid}{ID or ID string for a given KoboToolbox form}

\item{token}{optional KoboToolbox user access token}

\item{user}{optional KoboToolbox username}

\item{pass}{optional KoboToolbox password}
}
\value{
dataframe containing submitted data for the specified form
}
\description{
Get KoboToolbox form data as a dataframe for a given form
}
\details{
if \code{token} is supplied it is used preferentially over
\code{user} and \code{pass}. If token is not supplied, \code{user} and
\code{pass} must be supplied.
}
