% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboFormID.R
\name{koboFormID}
\alias{koboFormID}
\title{Get KoboToolbox form IDs for a user}
\usage{
koboFormID(server_url, token = NULL, user = NULL, pass = NULL)
}
\arguments{
\item{server_url}{URL of the KoboToolbox server,
e.g. \url{https://kc.kobotoolbox.org/}}

\item{token}{optional KoboToolbox user access token}

\item{user}{optional KoboToolbox username}

\item{pass}{optional KoboToolbox password}
}
\value{
dataframe containing form IDs and names for all forms accessible to
the specified user
}
\description{
Get KoboToolbox form IDs for a user
}
\details{
if \code{token} is supplied it is used preferentially over
\code{user} and \code{pass}. If token is not supplied, \code{user} and
\code{pass} must be supplied.
}
