% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latLong2UTM.R
\name{latLong2UTM}
\alias{latLong2UTM}
\title{Get valid UTM zone from latitude and longitude in WGS84 decimal degrees}
\usage{
latLong2UTM(x, y)
}
\arguments{
\item{x}{vector of longitude coordinates in decimal degrees}

\item{y}{vector of latitude coordinate in decimal degrees}
}
\value{
Vector of UTM zones for each latitude-longitude pair
}
\description{
Get valid UTM zone from latitude and longitude in WGS84 decimal degrees
}
