% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmAnova.R
\name{lmAnova}
\alias{lmAnova}
\title{Run an ANOVA on a list of linear models}
\usage{
lmAnova(mod_list)
}
\arguments{
\item{mod_list}{list of \code{lm()} model objects}
}
\description{
Run an ANOVA on a list of linear models
}
