% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFormat.R
\name{lmFormat}
\alias{lmFormat}
\title{Format beta coefficient of lm() model object for LaTeX text}
\usage{
lmFormat(x, digits = 2, ...)
}
\arguments{
\item{x}{lm() model object}

\item{digits}{number of digits to round F statistic}

\item{...}{arguments passed to \code{pFormat()}}
}
\value{
character string
}
\description{
Format beta coefficient of lm() model object for LaTeX text
}
\examples{
mod <- lm(rnorm(50) ~ rnorm(50))
lmFormat(mod)
}
