% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmPval.R
\name{lmPval}
\alias{lmPval}
\title{Get the model P-value for a linear model}
\usage{
lmPval(x)
}
\arguments{
\item{x}{lm model object}
}
\value{
p value of model
}
\description{
Get the model P-value for a linear model
}
