% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupTable.R
\name{lookupTable}
\alias{lookupTable}
\title{Use a lookup table to match values}
\usage{
lookupTable(x, y, old, new)
}
\arguments{
\item{x}{vector of original values}

\item{y}{dataframe lookup table}

\item{old}{column name string of original values}

\item{new}{column name string of new values}
}
\value{
vector of updated values
}
\description{
Use a lookup table to match values
}
\examples{
x <- c("A", "B", "B", "C", "A")
lookup <- data.frame(old = c("A", "B", "C", "D"), new = c(1, 2, 3, 4))
lookupTable(x, lookup, "old", "new")

}
