% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxChord.R
\name{maxChord}
\alias{maxChord}
\title{Find the longest chord within a polygon}
\usage{
maxChord(x)
}
\arguments{
\item{x}{st_POLYGON object}
}
\value{
st_LINESTRING
}
\description{
Find the longest chord within a polygon
}
\examples{
crd <- matrix(c(0, 0, 0.5, 1, 1, 0, 0, 1, 1.5, 1, 0, 0), ncol = 2)
poly <- st_polygon(list(crd))
maxChord(poly)


}
