% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxChord.R
\name{maxPerpChord}
\alias{maxPerpChord}
\title{Find the longest perpendicular chord within a polygon}
\usage{
maxPerpChord(x, chord = maxChord(x), theta = pi/2)
}
\arguments{
\item{x}{st_POLYGON}
}
\value{
st_LINESTRING
}
\description{
Find the longest perpendicular chord within a polygon
}
\examples{
crd <- matrix(c(0, 0, 0.5, 1, 1, 0, 0, 1, 1.5, 1, 0, 0), ncol = 2)
poly <- st_polygon(list(crd))
maxPerpChord(poly)
maxPerpChord(poly, theta = pi/3)

}
