% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minBox.R
\name{minBox}
\alias{minBox}
\title{Return the minimum bounding rectangle around a polygon}
\usage{
minBox(x)
}
\arguments{
\item{x}{sf object containing only geometries of type \code{POLYGON} or
\code{MULTIPOLYGON}}
}
\value{
list containing vertex points (\code{ptx}),
width (\code{width}) height (\code{height}) and angle of bounding
rectangle of each polygon.
}
\description{
Return the minimum bounding rectangle around a polygon
}
\examples{
dat <- sf::st_read(system.file("shape/nc.shp", package="sf"))
min_box_list <- minBox(dat)
min_box_list[[1]]

min_box_sf <- do.call(rbind, lapply(min_box_list, function(x) {
  pts_sf <- sf::st_as_sf(as.data.frame(x$pts), coords = c("X", "Y"))
  sf::st_sf(geometry = sf::st_convex_hull(sf::st_union(pts_sf)), crs = sf::st_crs(dat))
  }))
plot(sf::st_geometry(dat), col = NA, border = "blue")
plot(sf::st_geometry(min_box_sf), col = NA, border = "red", add = TRUE)

}
