% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMDSScree.R
\name{NMDSScree}
\alias{NMDSScree}
\title{Create scree plot for NMDS}
\usage{
NMDSScree(x, dims = 10, ...)
}
\arguments{
\item{x}{species by site abundance matrix}

\item{dims}{either an integer of the maximum number of dimensions or an
integer vector of dimensions to try}

\item{...}{additional arguments passed to \code{vegan::metaMDS()}}
}
\value{
plot object
}
\description{
Runs \code{vegan::metaMDS()} across a specified range of dimensions
}
