% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numFormat.R
\name{numFormat}
\alias{numFormat}
\title{Format a number for LaTeX}
\usage{
numFormat(x, digits = 2, method = "round")
}
\arguments{
\item{x}{atomic vector}

\item{digits}{number of digits}

\item{method}{method of rounding, either "round" or "signif"}
}
\value{
character vector
}
\description{
Format a number for LaTeX
}
