% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numWord.R
\name{numWord}
\alias{numWord}
\title{Convert numbers to English words}
\usage{
numWord(x, cap = FALSE, hyphen = TRUE, and = FALSE, round = FALSE)
}
\arguments{
\item{x}{integer vector}

\item{cap}{logical, if TRUE capitalise first letter of word.}

\item{hyphen}{logical, if TRUE insert \code{-} between 21 and
99 (except 30, 40, etc.).}

\item{and}{logical, if TRUE insert \code{and} between hundreds and tens,
e.g. 110 as \dQuote{one hundred and ten}}

\item{round}{logical, if TRUE decimals are rounded to integers,
using \code{round()}. If FALSE, decimals cause failure}
}
\value{
character vector.
}
\description{
Convert numbers to English words
}
\examples{
numWord(0, cap = TRUE)
numWord(0:121, and = TRUE)
numWord(1e6)
numWord(1e11+12345678)
numWord(-987654321)
numWord(1e15-1)
numWord(c(1,2,3,4.55), round = TRUE)
}
