% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objExist.R
\name{objExist}
\alias{objExist}
\title{Check if an object exists using the object itself, rather than char string}
\usage{
objExist(...)
}
\arguments{
\item{...}{any number of objects}
}
\value{
logical vector
}
\description{
Check if an object exists using the object itself, rather than char string
}
