% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect.R
\name{outersect}
\alias{outersect}
\title{Find values not shared between two vectors}
\usage{
outersect(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
vector of values not present in both vectors
}
\description{
Find values not shared between two vectors
}
\examples{
x <- c(1,2,3,4)
y <- c(1,3,5,7)
outersect(x, y)

}
