% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palShow.R
\name{palShow}
\alias{palShow}
\title{Show a colour palette as a plot}
\usage{
palShow(x)
}
\arguments{
\item{x}{vector of colour codes, either hex codes or string names}
}
\value{
Plot object
}
\description{
Show a colour palette as a plot
}
\examples{
palShow(c("#2166AC", "#B2182B"))

}
