% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteVals.R
\name{pasteVals}
\alias{pasteVals}
\title{Paste values together, replace NAs with blank, optional separator}
\usage{
pasteVals(
  ...,
  sep = "",
  collapse = NULL,
  remna = TRUE,
  unique = FALSE,
  sort = FALSE
)
}
\arguments{
\item{...}{vectors or dataframe to be pasted together}

\item{sep}{separator between adjacent values in vectors}

\item{collapse}{separator between sets of values across vectors}

\item{unique}{logical, if TRUE duplicated values are removed}

\item{sort}{logical, if TRUE values are sorted}
}
\value{
vector of pasted values
}
\description{
Paste values together, replace NAs with blank, optional separator
}
\examples{
pasteVals(c(1,2,3,4,5), c(5,4,3,2,1), sep = "|")
pasteVals(c(1,2,3,4,5), c(5,4,3,2,1), c("a", "b", "c", "d", "e"), 
  sep = "|", collapse = "-")
pasteVals(data.frame(c(1,2,3,4,5), c(5,4,3,2,1)))
pasteVals(c(1,2,3,NA_real_,NA_real_), c(5,4,NA_real_,2,NA_real_))
pasteVals(c(6,2,3,4,5,6), c(6,5,4,4,2,1), unique = TRUE)
pasteVals(c(6,2,3,4,5,6), c(6,5,4,4,NA_real_,1), sort = TRUE, remna = FALSE)
pasteVals(c(6,2,3,4,5,6), c(6,5,4,4,NA_real_,1), sort = TRUE)

}
