% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFormat.R
\name{pFormat}
\alias{pFormat}
\title{Format p-values for use in LaTeX}
\usage{
pFormat(
  p,
  lev = c(0.001, 0.01, 0.05),
  asterisks = FALSE,
  ps = TRUE,
  digits = NULL
)
}
\arguments{
\item{p}{numeric vector of p-values}

\item{lev}{vector of threshold values}

\item{asterisks, }{logical, alternatively return asterisks at thresholds}

\item{ps, }{logical, if TRUE return}

\item{digits}{optional, number of decimal places to round p values to if
not lower than any critical threshold}
}
\value{
character vector
}
\description{
Format p-values for use in LaTeX
}
\examples{
pFormat(0.03)
pFormat(c(0.06,0.05,0.005))
pFormat(0.02, ps = FALSE)
pFormat(0.02, asterisks = TRUE)
pFormat(0.55, digits = 1)
}
