% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playfair.R
\name{playfair}
\alias{playfair}
\title{Create a playfair-style cipher}
\usage{
playfair(x)
}
\arguments{
\item{x}{character string to encode}
}
\value{
list with three slots: (1) encoded message (2) decoder matrix
(3) decoder lookup table
}
\description{
Create a playfair-style cipher
}
\details{
Creates a cipher based on the original playfair cipher.
Unlike the original playfair cipher this method produces a
6x6 grid of upper and lowercase letters. Additionally, the
behaviour when a keypair appear on the same row or column of
the decoder matrix is different. In this version keypairs which
appear on the same row or column are merely swapped rather than
transposed as in the original cipher.
Messages to be encoded are converted to uppercase and all
non-alphabet characters are stripped out.
}
\examples{
x <- "This is a test"
playfair(x)

}
