% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmFormat.R
\name{pm_format}
\alias{pm_format}
\title{Format a plus-minus number combo}
\usage{
pm_format(x, y, dx = 2, dy = dx + 1, pm = "$\\\\pm$", paren = FALSE)
}
\arguments{
\item{x}{atomic vector}

\item{y}{atomic vector}

\item{dx}{number of decimal places of x}

\item{dy}{number of decimal places of y}

\item{pm}{LaTeX plus-minus symbol}

\item{paren}{logical, should the measure of uncertainty be placed in
parentheses?}
}
\value{
character string
}
\description{
Format a plus-minus number combo
}
