% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyVert.R
\name{polyVert}
\alias{polyVert}
\title{Generate regular 2D shapes by defining the number of sides and the radius}
\usage{
polyVert(nsides, radius, centre = c(0, 0), angle = 0)
}
\arguments{
\item{nsides}{number of sides/vertices}

\item{radius}{radius of shape}

\item{centre}{numeric vector of length 2, defining the x,y coordinates of the shape centre}

\item{angle}{starting angle for first vertex}
}
\value{
dataframe of vertices bounding the shape
}
\description{
Generate regular 2D shapes by defining the number of sides and the radius
}
\examples{
polyVert(5, 10)
polyVert(4, 5, centre = c(10, 1), angle = 45)

}
