% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadMean.R
\name{quadMean}
\alias{quadMean}
\title{Calculate the quadratic mean}
\usage{
quadMean(x)
}
\arguments{
\item{x}{vector}
}
\value{
quadratic mean of vector
}
\description{
Calculate the quadratic mean
}
\examples{
x <- seq(0, 10)
quadMean(x)
}
