% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualPal.R
\name{qualPal}
\alias{qualPal}
\title{Qualitative colour palettes}
\usage{
qualPal()
}
\value{
named list of qualitative colour palettes as vectors of hex-codes
}
\description{
Qualitative colour palettes
}
\details{
\code{color_brewer_*} come from \url{https://colorbrewer2.org/}.
\code{seaborn_*} come from \url{https://seaborn.pydata.org/}.
\code{carto_colors_*} come from \url{https://carto.com/carto-colors/}.
\code{solarized_accent} comes from \url{https://ethanschoonover.com/solarized/}.
\code{coloropt_*} come from \url{http://tsitsul.in/blog/coloropt/}.
\code{rainbow_12bit} comes from \url{https://iamkate.com/data/12-bit-rainbow/}.
\code{oksolar} comes from \url{https://meat.io/oksolar}.
Other palettes are original.
}
\examples{
qualPal()[[1]]

}
