% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastFort.R
\name{rastFort}
\alias{rastFort}
\title{Convert a raster to a dataframe for plotting with ggplot2}
\usage{
rastFort(x)
}
\arguments{
\item{x}{raster object}
}
\value{
dataframe object
}
\description{
Convert a raster to a dataframe for plotting with ggplot2
}
