% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangleBuffer.R
\name{rectangleBuffer}
\alias{rectangleBuffer}
\title{Create a rectangular buffer around an sfc point object}
\usage{
rectangleBuffer(x, x_len, y_len, deg = 0)
}
\arguments{
\item{x}{sfc containing points}

\item{x_len}{length of rectangle along x axis}

\item{y_len}{length of rectangle along y axis}

\item{deg}{angle in degrees of rotation}
}
\value{
sf polygon object
}
\description{
Create a rectangular buffer around an sfc point object
}
