% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remDupString.R
\name{remDupSring}
\alias{remDupSring}
\title{Remove duplicated characters within strings}
\usage{
remDupSring(x)
}
\arguments{
\item{x}{vector of character strings}
}
\value{
vector of character strings with duplicated values removed
}
\description{
Remove duplicated characters within strings
}
\examples{
remDupFlag(c("FU", "FUFF", "FJJ"))


}
