% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateMatrix.R
\name{rotate}
\alias{rotate}
\title{Rotate a matrix 90 degrees clockwise or anti-clockwise}
\usage{
rotate(x, clockwise = TRUE)
}
\arguments{
\item{x}{matrix}

\item{clockwise}{logical, if TRUE rotate clockwise, if FALSE rotate
anti-clockwise}
}
\value{
matrix
}
\description{
Rotate a matrix 90 degrees clockwise or anti-clockwise
}
