% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundAny.R
\name{roundAny}
\alias{roundAny}
\title{Round a number to nearest multiple of a given number}
\usage{
roundAny(x, y, fun = round)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Multiple to round to}

\item{fun}{function for rounding, either \code{round}, \code{floor},
\code{ceiling}}
}
\value{
Vector of rounded values
}
\description{
Round a number to nearest multiple of a given number
}
\examples{
a <- seq(1, 20, 0.5)
roundAny(a, 2)
roundAny(a, 2, floor)

}
