% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundMultiple.R
\name{roundMultiple}
\alias{roundMultiple}
\title{Round a number to nearest multiple of a given number}
\usage{
roundMultiple(x, y)
}
\arguments{
\item{x}{Numveric vector}

\item{y}{Multiple to round to}
}
\value{
Vector of rounded values
}
\description{
Round a number to nearest multiple of a given number
}
\examples{
a <- seq(1, 20, 0.5)
roundMultiple(a, 2)
}
