% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale01.R
\name{scale01}
\alias{scale01}
\title{Scale a vector to between zero and one}
\usage{
scale01(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Numeric vector with values scaled between zero and one
}
\description{
Scale a vector to between zero and one
}
\examples{
scale01(1:1000)
 
}
