% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilQuery.R
\name{soilQuery}
\alias{soilQuery}
\title{Query SoilGrids via the REST API}
\usage{
soilQuery(
  x,
  plot_id = "plot_id",
  longitude_of_centre = "longitude_of_centre",
  latitude_of_centre = "latitude_of_centre",
  attrib = c("cec", "cfvo", "clay", "nitrogen", "ocd", "ocs", "phh20", "sand", "silt",
    "soc"),
  depth = c("0-5", "5-15", "15-30", "30-60", "60-100", "100-200"),
  average = list(c(0, 30)),
  value = "mean"
)
}
\arguments{
\item{x}{dataframe of plot level data}

\item{plot_id}{column name string of plot IDs}

\item{longitude_of_centre}{column name string of plot longitude}

\item{latitude_of_centre}{column name string of plot latitude}

\item{attrib}{vector of soil attributes to extract}

\item{depth}{vector of soil depths over which to extract the attributes, e.g. 0-5, 30-60}

\item{average}{optional list of vectors, each of length 2, describing ranges
of depths over which to average (mean) the values of attrib}

\item{value}{vector of soil values to extract for each attribute, e.g. mean, Q0.05}
}
\value{
dataframe of soil values by depth
}
\description{
See https://rest.soilgrids.org/soilgrids/v2.0/docs for more information
}
\details{
See \code{soilQueryOptions()} for all possibly combinations of
attrib, depth, and value
}
