% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdErrBinom.R
\name{stdErrBinom}
\alias{stdErrBinom}
\title{Calculate standard error on binomial data}
\usage{
stdErrBinom(x)
}
\arguments{
\item{x}{TRUE FALSE vector}
}
\value{
standard error
}
\description{
Calculate standard error on binomial data
}
\examples{
a <- c(TRUE, FALSE, TRUE, FALSE, TRUE, TRUE)
stdErrBinom(a)
}
