% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdErrMean.R
\name{stdErrMean}
\alias{stdErrMean}
\title{Calculate the standard error of the mean of a vector}
\usage{
stdErrMean(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Standard error of the mean
}
\description{
Calculate the standard error of the mean of a vector
}
\examples{
x <- rnorm(100)
stdErrMean(x)

}
