% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texCmd.R
\name{texCmd}
\alias{texCmd}
\title{Generate a LaTeX command from R object}
\usage{
texCmd(stat, func)
}
\arguments{
\item{stat}{atomic vector}

\item{func}{character string with name of LaTeX function}
}
\value{
character string
}
\description{
Generate a LaTeX command from R object
}
\examples{
a <- 3.145
texCmd(a, "pi")
}
