% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themePub.R
\name{themePub}
\alias{themePub}
\title{A sensible theme for basic ggplot output for use in LaTeX articles}
\usage{
themePub(
  base_size = 12,
  base_family = "serif",
  base_line_size = base_size/22,
  base_rect_size = base_line_size,
  ...
)
}
\arguments{
\item{base_size}{base font size, given in pts}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{...}{further arguments passed to \code{ggplot2::theme()}}
}
\description{
A sensible theme for basic ggplot output for use in LaTeX articles
}
\examples{
data(mtcars)
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = wt)) + 
 geom_point() + 
 themePub()
}
