% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usingPkg.R
\name{usingPkg}
\alias{usingPkg}
\title{Dynamically install packages as needed and load them}
\usage{
usingPkg(...)
}
\arguments{
\item{...}{packages as character strings}
}
\description{
Dynamically install packages as needed and load them
}
\examples{
usingPkg("dplyr", "tidyr", "ggplot2")
}
