% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTMProj4.R
\name{UTMProj4}
\alias{UTMProj4}
\title{Generate a valid UTM WGS84 proj4string given a UTM zone}
\usage{
UTMProj4(x)
}
\arguments{
\item{x}{character vector defining UTM zones}
}
\value{
UTM proj4string character vector
}
\description{
Generate a valid UTM WGS84 proj4string given a UTM zone
}
