% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnFormat.R
\name{warnFormat}
\alias{warnFormat}
\title{Format warnings, messages and errors}
\usage{
warnFormat(x, warn, type, n = 10, sep = ",")
}
\arguments{
\item{x}{vector of row positions, IDs etc. to display}

\item{warn}{character string with warning, message or error}

\item{type}{choose either "warning", "message" or "error"}

\item{n}{integer, number of items from x to return in warning}
}
\description{
Format warnings, messages and errors
}
