% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtableHline.R
\name{xtableHline}
\alias{xtableHline}
\title{Calculate vector of hline positions for xtable's hline.after parameter}
\usage{
xtableHline(ngroup, nsize)
}
\arguments{
\item{ngroup}{number of groups in the table}

\item{nsize}{number of members in each group}
}
\value{
vector of hline positions
}
\description{
Calculate vector of hline positions for xtable's hline.after parameter
}
\examples{
xtableHline(5,4)
xtableHline(3,c(4,3,2))

}
