% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestMatch.R
\name{closestMatch}
\alias{closestMatch}
\title{Find closest match in a vector}
\usage{
closestMatch(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{vector of numeric values, each of which be matched with the closest
value in \code{x}}
}
\value{
numeric vector of values of \code{x} that are closest to each
element of \code{y}
}
\description{
Find closest match in a vector
}
