% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFormat.R
\name{pFormat}
\alias{pFormat}
\title{Format p-values for use in LaTeX}
\usage{
pFormat(
  p,
  lev = c(0.001, 0.01, 0.05),
  round = TRUE,
  digits = 2,
  asterisks = FALSE
)
}
\arguments{
\item{p}{numeric vector of p-values}

\item{lev}{vector of threshold values}

\item{round}{logical, if TRUE values not lower than any critical threshold
are rounded to \code{n} decimal places}

\item{digits}{number of decimal places to round p values to if not lower than
any critical threshold}

\item{asterisks, }{logical, alternatively return asterisks at thresholds}
}
\value{
character vector
}
\description{
Format p-values for use in LaTeX
}
\examples{
pFormat(0.03)
pFormat(c(0.06,0.05,0.005))
}
