% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualPal.R
\name{qualPal}
\alias{qualPal}
\title{Qualitative colour palettes}
\usage{
qualPal()
}
\value{
named list of qualitative colour palettes as vectors of hex-codes
}
\description{
Qualitative colour palettes
}
\details{
The \code{color_brewer_*} palettes come from \url{https://colorbrewer2.org/}.
The \code{seaborn_*} palettes come from \url{https://seaborn.pydata.org/}.
The \code{carto_colors_*} palettes come from \url{https://carto.com/carto-colors/}.
The \code{solarized_accent} palette comes from \url{https://ethanschoonover.com/solarized/}.
The \code{coloropt_*} palettes come from \url{http://tsitsul.in/blog/coloropt/}.
Other palettes are original.
}
\examples{
qualPal()[[1]]

}
